% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artists.df}
\alias{artists.df}
\title{artists.df}
\format{
An object of class \code{data.frame} with 18708 rows and 1 columns.
}
\usage{
artists.df
}
\description{
Names of artists
}
\details{
@format ## `artists.df`
 A data frame with 18708 rows and 1 column:
 \describe{
   \item{Artist}{Name of the artists}
 }
 @source <https://genius.com/artists>
}
\keyword{datasets}
