% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{compare.lyrics}
\alias{compare.lyrics}
\title{Compare lyrics}
\usage{
compare.lyrics(name.x, name.y, language = "en")
}
\arguments{
\item{name.x}{The name of the first artist of your choice.}

\item{name.y}{The name of the second artist of your choice.}

\item{language}{The language of the lyrics you want.}
}
\value{
returns a plot that compares the lyrics between the two artists.
}
\description{
Compare lyrics
}
\examples{
\dontrun{
compare.lyrics("lady gaga", "lana del rey")
}
}
\author{
Emile Paris
}
