% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{compare.year}
\alias{compare.year}
\title{Compare year}
\usage{
compare.year(name.x, name.y, local = TRUE)
}
\arguments{
\item{name.x}{The name of the first artist of your choice.}

\item{name.y}{The name of the second artist of your choice.}

\item{local}{determine if the function is run locally or not.}
}
\value{
return a density plot of the year of activity & popularity of the two artists.
}
\description{
Compare year
}
\examples{
\dontrun{
compare.year("lady gaga", "lana del rey", local = TRUE)
}
}
\author{
Emile Paris
}
