% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{Geniusdashboard}
\alias{Geniusdashboard}
\title{Geniusdashboard}
\usage{
Geniusdashboard(artists, lang)
}
\arguments{
\item{artists}{a dataframe with one column containing the artists you want.}

\item{lang}{a dataframe with one column containing the abrevations of the languages you want (example: English must be "en").}
}
\value{
returns a dashboard
}
\description{
Geniusdashboard
}
\examples{
\dontrun{
Geniusdashboard(artists.df, lang.df)
}
}
\author{
Emile Paris
}
