% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lang.df}
\alias{lang.df}
\title{lang.df}
\format{
An object of class \code{data.frame} with 57 rows and 1 columns.
}
\usage{
lang.df
}
\description{
Abreviations of languages
}
\details{
@format ## `lang.df`
 A data frame with 57 rows and 1 column:
 \describe{
   \item{stopwords::stopwords_getlanguages("stopwords-iso")}{Abreviations of languages}
 }
 @source <https://genius.com/artists>
}
\keyword{datasets}
