% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{TopAlbums}
\alias{TopAlbums}
\title{Top Albums}
\usage{
TopAlbums(name)
}
\arguments{
\item{name}{The name of the artist of your choice.}
}
\value{
returns the albums and the release date of your artist.
}
\description{
Top Albums
}
\examples{
TopAlbums("Lana Del Rey")
}
\author{
Emile Paris
}
