% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{TopAlbums}
\alias{TopAlbums}
\title{Top Albums}
\usage{
TopAlbums(name, local = TRUE)
}
\arguments{
\item{name}{The name of the artist of your choice.}

\item{local}{determine if the function is run locally or not.}
}
\value{
Return the name of the albums and the year of release of a specific artist (name).
}
\description{
Top Albums
}
\examples{
\dontrun{
TopAlbums("Lana Del Rey", local = TRUE)
}
}
\author{
Emile Paris
}
