% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{TopAlbums}
\alias{TopAlbums}
\title{Top Albums}
\usage{
TopAlbums(name, local = TRUE)
}
\arguments{
\item{name}{The name of the artist of your choice.}

\item{local}{determine if the function is run locally or not.}
}
\value{
returns the albums and the release date of your artist.
}
\description{
Top Albums
}
\examples{
TopAlbums("Lana Del Rey", local = TRUE)
}
\author{
Emile Paris
}
