% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{Topsongslyrics}
\alias{Topsongslyrics}
\title{Top songs lyrics}
\usage{
Topsongslyrics(
  name,
  tfgraph = FALSE,
  wordcloud = FALSE,
  lang,
  environment.return = TRUE,
  local = TRUE
)
}
\arguments{
\item{name}{The name of the artist of your choice.}

\item{tfgraph}{If True, returns a term-frequency graph.}

\item{wordcloud}{If True, returns a worldcloud.}

\item{lang}{Language of the lyrics.}

\item{environment.return}{If True, saves all computations in global environment.}

\item{local}{determine if the function is run locally or not.}
}
\value{
returns a dataframe with the 10 top songs and the lyrics for a given  artist (name) speaking in a special language (lang). Possibility to print a Term frequency graph and a wordcloud. Also, possibility to return directly in the environment the df, token, corpus, dtm, freq, and covariance matrix.
}
\description{
Top songs lyrics
}
\examples{
\dontrun{
Topsongslyrics(name = "Lady Gaga",
               tfgraph = TRUE,
               wordcloud = TRUE,
               lang = "en",
               environment.return = TRUE,
               local = TRUE)
}
}
\author{
Emile Paris
}
