% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{Topsongslyrics}
\alias{Topsongslyrics}
\title{Top songs lyrics}
\usage{
Topsongslyrics(
  name,
  tfgraph = FALSE,
  wordcloud = FALSE,
  lang,
  environment.return = TRUE
)
}
\arguments{
\item{name}{The name of the artist of your choice.}

\item{tfgraph}{If True, returns a term-frequency graph.}

\item{lang}{Language of the lyrics.}

\item{environment.return}{If True, saves all computations in global environment.}

\item{worldcloud}{If True, returns a worldcloud.}
}
\value{
returns the top 10 songs of your artist with the lyrics.
}
\description{
Top songs lyrics
}
\examples{
Topsongslyrics(name = "Lady Gaga", tfgraph = FALSE, wordcloud = FALSE, lang = "en", environment.return = TRUE)
}
\author{
Emile Paris
}
